+++ src/app/stats/components/ListeningClockComponent.tsx
@@ -2,77 +2,26 @@
 import React, { useEffect, useRef } from 'react';
 import * as d3 from 'd3';
 import { months } from '@/util/dateTimeFormat';
+import { NumberByMonth } from '@/types';
 
 interface ListeningClockComponentProps {
-    fileContent: string;
-    selectedArtist: string;
-    startDate: string;
-    endDate: string;
+    data: NumberByMonth[];
+    year: string;
 }
 
-interface ListeningTimeByMonth {
-    month: string;
-    value: number;
-}
-
-/**
- * This function gets the listening time for each month of the year for the selected artist
- * @param fileContent the JSON content of the file
- * @param selectedArtist the artist selected by the user
- * @returns an array of objects with the month and the total listening time for that month {
- * month: string,
- * value: number}
- */
-// ListeningClockComponent.tsx
-
-const getListeningTimeByMonth = (fileContent: string, selectedArtist: string, startDate: string, endDate: string): ListeningTimeByMonth[] => {
-    const data = JSON.parse(fileContent);
-
-    const listeningTimeByMonth = new Map<string, number>();
-
-    data.forEach((record: { artistName: string; endTime: string; msPlayed: number }) => {
-        // Convert record's endTime to YYYY-MM-DD format
-        const recordTime = record.endTime.split(' ')[0];
-
-        // Only process records within the date range
-        if (recordTime >= startDate && recordTime <= endDate) {
-            const key = record.endTime.split('-')[1];
-            if (record.artistName === selectedArtist) {
-                const currentListeningTime = listeningTimeByMonth.get(key) || 0;
-                listeningTimeByMonth.set(key, currentListeningTime + record.msPlayed);
-            }
-        }
-    });
-
-    const listeningTimeByMonthArray = Array.from(listeningTimeByMonth).map(([key, value]) => ({
-        month: key,
-        value: Number((value / 60000).toFixed(1)),
-    }));
-
-    // for months with no listening time, add 0
-    for (let i = 1; i <= 12; i++) {
-        if (!listeningTimeByMonthArray.some((record: ListeningTimeByMonth) => record.month === i.toString())) {
-            listeningTimeByMonthArray.push({ month: i.toString(), value: 0 });
-        }
-    }
-
-    return listeningTimeByMonthArray.sort((a, b) => Number(a.month) - Number(b.month));
-};
 
 
-const ListeningClockComponent: React.FC<ListeningClockComponentProps> = ({ fileContent, selectedArtist, startDate, endDate }) => {
+const ListeningClockComponent: React.FC<ListeningClockComponentProps> = ({ data, year }) => {
   const ref = useRef(null);
-    
-  const data = getListeningTimeByMonth(fileContent, selectedArtist, startDate, endDate);
-
+  const width = 500; // Adjust as needed
+  const height = 500; // Adjust as needed
   useEffect(() => {
     if (data.length === 0) return;
     // Remove previous chart
     d3.select(ref.current).selectAll('*').remove();
 
     const svg = d3.select(ref.current);
-    const width = 300; // Adjust as needed
-    const height = 300; // Adjust as needed
+    
     const radius = Math.min(width, height) / 2;
     
     // Scales for the bars
@@ -123,7 +72,7 @@ const ListeningClockComponent: React.FC<ListeningClockComponentProps> = ({ fileC
 
   return (
     <div className="bg-dark flex justify-center h-full w-full flex-grow">
-        <svg ref={ref} width={500} height={500} />
+        <svg ref={ref} width={width} height={height} />
     </div>
 
   );
